added the concept of mini chunks
mini chunks are cubes with a side length of 16 inside of a chunk.
They contain a list of all cells present within that cube. This way, when searching for the cell filling a certain position, you dont have to go through every cell in that chunk, but only through every cell within a mini chunk.